order的合并问题暂时解决

+++ src/main/java/com/jbgz/pancakejob/entity/Order.java
@@ -55,12 +55,12 @@ public class Order implements Serializable {
     /**
      * 
      */
-    private BigDecimal jobhunterScore;
+    private Integer jobhunterScore;
 
     /**
      * 
      */
-    private BigDecimal recruiterScore;
+    private Integer recruiterScore;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
